DATAMONGO-1643 - Remove unused ReflectiveDBCollectionInvoker.
